Decided to upload crypto wrapper/helper js files in addition to other files.
Rationale behind this is:
1) This is not meant to be a cryptographically secure extension; that is, outside parties still have influence
over the communication of Alice and Bob who both have the extension.
2) This code will be in the compiled JavaScript files anyways, so hiding it here is merely a matter of mild
obfuscation.
3) The combination of 1) and 2) implies that even by hiding these files, a third party can, without too much
difficulty, tamper with the communication of Alice and Bob to force the extension to fail, falsely authenticate,
block certain messages, etc. While it is unlikely that this would happen, there is not really any clear benefit
to keeping the source code private.
4) Version control will be very helpful for more sensitive files such as the crypto-related wrapper and helper
objects/functions.

The combination of minimal drawbacks and significant benefit (version control in the case of data loss,
regressions in the code, etc) makes it worth putting the files on GitHub.